Qualifications without responses were being discarded instead of warning the user.

+++ app/models/participation.rb
@@ -22,7 +22,7 @@ class Participation < ActiveRecord::Base
     survey.questions.each do |question|
       attributes = question_response_params[question.id.to_s] || {}
       current_responses = response[question.id]
-      if attributes[:response].blank? then
+      if attributes[:response].blank? && attributes[:qualifications].blank? then
         # didn't respond
         response[question.id].first.destroy if response[question.id] # So, destroy the previous response if it exists
       else
